Added basic testing for CI (#17)
* Added basic testing and implemented Travis CI
