+++ .gitignore
@@ -4,5 +4,5 @@ routes/
 vendor/*/
 .idea
 .DS_Store
-node_modules
+node_modules/
 log/
