+++ .travis.yml
@@ -1,0 +1,31 @@
+language: go
+
+node_js:
+  - 8
+
+go:
+  - 1.9
+
+os:
+  - osx
+  - linux
+
+cache:
+  directories:
+    - ~/.npm
+    - ~/.cache
+
+before_install: if [[ `npm -v` != 5.7* ]]; then npm i -g npm@5; fi
+
+install:
+  - go get -v github.com/revel/cmd/revel
+  - go get -v github.com/kardianos/govendor
+  - npm ci
+  - govendor sync
+
+before_script:
+  - revel run github.com/anikitenko/mini-sftp-client &
+
+script:
+  - $(npm bin)/cypress run --record
+  - kill $(jobs -p) || true
