+++ app/views/footer.html
@@ -2,7 +2,8 @@
     {{template "debug.html" .}}
   {{end}}
 <script src="/public/js/main.js"></script>
-<script src="/public/js/navigation.js"></script>
+<script src="/public/js/navigation-remote.js"></script>
+<script src="/public/js/navigation-local.js"></script>
 <script src="/public/js/files.js"></script>
 </body>
 </html>
