+++ app/views/header.html
@@ -11,28 +11,23 @@
 
     <link rel="shortcut icon" type="image/png" href="/public/img/favicon.png">
 
-    <link rel="stylesheet" type="text/css"
-          href="https://cdnjs.cloudflare.com/ajax/libs/twitter-bootstrap/3.3.7/css/bootstrap.min.css">
-    <link rel="stylesheet" type="text/css"
-          href="https://cdnjs.cloudflare.com/ajax/libs/x-editable/1.5.1/bootstrap3-editable/css/bootstrap-editable.css">
-    <link rel="stylesheet" type="text/css"
-          href="https://cdnjs.cloudflare.com/ajax/libs/animate.css/3.5.2/animate.min.css">
-    <link rel="stylesheet" type="text/css"
-          href="https://cdnjs.cloudflare.com/ajax/libs/Ladda/1.0.5/ladda-themeless.min.css">
-    <link rel="stylesheet" type="text/css"
-          href="https://cdnjs.cloudflare.com/ajax/libs/select2/4.0.5/css/select2.min.css">
-    <link rel="stylesheet" type="text/css" href="/public/css/jquery-ui.min.css">
+    <link rel="stylesheet" type="text/css" href="/public/css/lib/bootstrap/bootstrap.min.css">
+    <link rel="stylesheet" type="text/css" href="/public/css/lib/bootstrap/bootstrap-editable.css">
+    <link rel="stylesheet" type="text/css" href="/public/css/lib/animate/animate.min.css">
+    <link rel="stylesheet" type="text/css" href="/public/css/lib/ladda/ladda-themeless.min.css">
+    <link rel="stylesheet" type="text/css" href="/public/css/lib/select2/select2.min.css">
+    <link rel="stylesheet" type="text/css" href="/public/css/lib/jquery/jquery-ui.min.css">
     <link rel="stylesheet" type="text/css" href="/public/css/style.css">
 
-    <script src="https://cdnjs.cloudflare.com/ajax/libs/jquery/3.2.1/jquery.min.js"></script>
-    <script src="https://cdnjs.cloudflare.com/ajax/libs/twitter-bootstrap/3.3.7/js/bootstrap.min.js"></script>
-    <script src="https://cdnjs.cloudflare.com/ajax/libs/x-editable/1.5.1/bootstrap3-editable/js/bootstrap-editable.min.js"></script>
-    <script src="https://cdnjs.cloudflare.com/ajax/libs/bootstrap-show-password/1.0.3/bootstrap-show-password.min.js"></script>
-    <script src="/public/js/bootstrap-notify.js"></script>
-    <script src="https://cdnjs.cloudflare.com/ajax/libs/Ladda/1.0.5/spin.min.js"></script>
-    <script src="https://cdnjs.cloudflare.com/ajax/libs/Ladda/1.0.5/ladda.min.js"></script>
-    <script src="https://cdnjs.cloudflare.com/ajax/libs/select2/4.0.5/js/select2.min.js"></script>
-    <script src="/public/js/jquery-ui.min.js"></script>
-    <script src="https://cdnjs.cloudflare.com/ajax/libs/jqueryui-touch-punch/0.2.3/jquery.ui.touch-punch.min.js"></script>
+    <script src="/public/js/lib/jquery/jquery.min.js"></script>
+    <script src="/public/js/lib/bootstrap/bootstrap.min.js"></script>
+    <script src="/public/js/lib/bootstrap/bootstrap-editable.min.js"></script>
+    <script src="/public/js/lib/bootstrap/bootstrap-show-password.min.js"></script>
+    <script src="/public/js/lib/bootstrap/bootstrap-notify.js"></script>
+    <script src="/public/js/lib/spin/spin.min.js"></script>
+    <script src="/public/js/lib/ladda/ladda.min.js"></script>
+    <script src="/public/js/lib/select2/select2.min.js"></script>
+    <script src="/public/js/lib/jquery/jquery-ui.min.js"></script>
+    <script src="/public/js/lib/jquery/jquery.ui.touch-punch.min.js"></script>
 </head>
 <body class="grey">
