+++ public/js/lib/bootstrap/bootstrap-notify.js
